Updated multiple types to use macros to generate constructors and getters.
