+++ deep_causality/src/types/context_types/node_types/dateoid.rs
@@ -4,28 +4,19 @@
 
 use std::fmt::{Display, Formatter};
 
+use deep_causality_macros::{Constructor, Getters};
+
 use crate::prelude::{Datable, Identifiable};
 
+#[derive(Constructor, Getters)]
 #[derive(Debug, Copy, Clone, Hash, Eq, PartialEq)]
 pub struct Dataoid
 {
+    #[getter(name = data_id)] // Rename ID getter to prevent conflict impl with identifiable
     id: u64,
     data: i32,
 }
 
-
-impl Dataoid
-{
-    pub fn new(id: u64, data_range: i32) -> Self {
-        Self { id, data: data_range }
-    }
-    pub fn data(&self) -> i32 {
-        self.data
-    }
-}
-
-// Optional. Override only when needed.
-// impl Adjustable for Dataoid {}
 
 impl Datable for Dataoid {}
