+++ deep_causality/src/types/context_types/node_types/root.rs
@@ -2,18 +2,14 @@
 // Copyright (c) "2023" . The DeepCausality Authors. All Rights Reserved.
 use std::fmt::{Display, Formatter};
 
+use deep_causality_macros::Constructor;
+
 use crate::protocols::identifiable::Identifiable;
 
+#[derive(Constructor)]
 #[derive(Debug, Copy, Clone, Hash, Eq, PartialEq)]
 pub struct Root {
     id: u64,
-}
-
-impl Root
-{
-    pub fn new(id: u64) -> Self {
-        Self { id }
-    }
 }
 
 impl Identifiable for Root
