+++ deep_causality/src/types/context_types/node_types/spaceoid.rs
@@ -3,8 +3,11 @@
 
 use std::fmt::{Display, Formatter};
 
+use deep_causality_macros::Constructor;
+
 use crate::prelude::{Identifiable, Spatial};
 
+#[derive(Constructor)]
 #[derive(Debug, Copy, Clone, Hash, Eq, PartialEq)]
 pub struct Spaceoid
 {
@@ -13,16 +16,6 @@ pub struct Spaceoid
     y: i64,
     z: i64,
 }
-
-impl Spaceoid
-{
-    pub fn new(id: u64, x: i64, y: i64, z: i64) -> Self {
-        Self { id, x, y, z }
-    }
-}
-
-// Optional. Override only when needed.
-// impl Adjustable for Spaceoid {}
 
 
 impl Identifiable for Spaceoid
@@ -52,10 +45,7 @@ impl Spatial for Spaceoid
 impl Display for Spaceoid {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "Spaceoid: id={}, x={}, y={}, z={}",
-               self.id,
-               self.x,
-               self.y,
-               self.z
+               self.id, self.x, self.y, self.z
         )
     }
 }
