+++ deep_causality/src/types/context_types/node_types/tempoid.rs
@@ -3,8 +3,11 @@
 
 use std::fmt::Display;
 
+use deep_causality_macros::Constructor;
+
 use crate::prelude::{Identifiable, Temporable, Temporal, TimeScale};
 
+#[derive(Constructor)]
 #[derive(Debug, Copy, Clone, Hash, Eq, PartialEq)]
 pub struct Tempoid
 {
@@ -13,25 +16,12 @@ pub struct Tempoid
     time_unit: u32,
 }
 
-impl Tempoid
-{
-    pub fn new(id: u64, time_scale: TimeScale, time_unit: u32) -> Self {
-        Self {
-            id,
-            time_scale,
-            time_unit,
-        }
-    }
-}
 
 impl Identifiable for Tempoid {
     fn id(&self) -> u64 {
         self.id
     }
 }
-
-// Optional. Override only when needed.
-// impl Adjustable for Tempoid {}
 
 impl Temporal for Tempoid {}
