+++ deep_causality/src/types/csm_types/csm_action.rs
@@ -1,28 +1,16 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) "2023" . The DeepCausality Authors. All Rights Reserved.
 
+use deep_causality_macros::{Constructor, Getters};
+
 use crate::prelude::ActionError;
 
+#[derive(Getters, Constructor)]
 #[derive(Clone, Debug)]
 pub struct CausalAction {
     action: fn() -> Result<(), ActionError>,
     descr: &'static str,
     version: usize,
-}
-
-impl CausalAction
-{
-    pub fn new(action: fn() -> Result<(), ActionError>, descr: &'static str, version: usize) -> Self {
-        Self { action, descr, version }
-    }
-
-    pub fn descr(&self) -> &'static str {
-        self.descr
-    }
-
-    pub fn version(&self) -> usize {
-        self.version
-    }
 }
 
 impl CausalAction
