+++ deep_causality/src/types/reasoning_types/inference/mod.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) "2023" . The DeepCausality Authors. All Rights Reserved.
 
+use deep_causality_macros::Constructor;
+
 use crate::prelude::{DescriptionValue, IdentificationValue, NumericalValue};
 
 mod identifiable;
@@ -8,6 +10,7 @@ mod inferable;
 mod display;
 
 
+#[derive(Constructor)]
 #[derive(Debug, Clone, PartialEq, PartialOrd)]
 pub struct Inference
 {
@@ -17,16 +20,4 @@ pub struct Inference
     threshold: NumericalValue,
     effect: NumericalValue,
     target: NumericalValue,
-}
-
-impl Inference
-{
-    pub fn new(
-        id: IdentificationValue, question: DescriptionValue, observation: NumericalValue,
-        threshold: NumericalValue, effect: NumericalValue, target: NumericalValue,
-    )
-        -> Self
-    {
-        Self { id, question, observation, threshold, effect, target }
-    }
 }
